smokey/posix-mutex: Fix test case by destroying barriers correctly
At least on x86-64-compat, the missing destruction of the smokey
barriers, specifically their embedded mutexes, cause crashes of the
test. The reason is likely a mismatch between the kernel's and
userland's view on which object is still active, combined with the fact
that userland kept them on the volatile stack.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>